＜統合テストの実装＞ ・基本認証機能、管理者専用機能、不動産機能についてコントローラーからH2DBまでの統合テストを実装 ・統合テストに合わせてH2DBのスキーマと初期データを変更 ・初期データのユーザーIDの型をMySQLに合わせてBIGINTからINTに修正 ・H2DBをインメモリで起動するようにパラメーターを変更 ・統合テストでトークン生成が必要となる為JWT関係のパラメーターを追記

+++ src/test/resources/schema.sql
@@ -1,13 +1,13 @@
-CREATE TABLE IF NOT EXISTS projects(id INT PRIMARY KEY AUTO_INCREMENT, user_id BIGINT NOT NULL, project_name VARCHAR(100),is_deleted boolean);
+CREATE TABLE IF NOT EXISTS projects(id INT PRIMARY KEY AUTO_INCREMENT, user_id INT NOT NULL, project_name VARCHAR(100),is_deleted boolean);
 
-CREATE TABLE IF NOT EXISTS parcels(id INT PRIMARY KEY AUTO_INCREMENT,project_id INT NOT NULL, user_id BIGINT NOT NULL, parcel_price INT,parcel_address VARCHAR(100),parcel_category VARCHAR(50),parcel_size DOUBLE,parcel_remark VARCHAR(100),is_deleted boolean);
+CREATE TABLE IF NOT EXISTS parcels(id INT PRIMARY KEY AUTO_INCREMENT,project_id INT NOT NULL, user_id INT NOT NULL, parcel_price BIGINT,parcel_address VARCHAR(100),parcel_category VARCHAR(50),parcel_size DOUBLE,parcel_remark VARCHAR(100),is_deleted boolean);
 
-CREATE TABLE IF NOT EXISTS buildings(id INT PRIMARY KEY AUTO_INCREMENT, project_id INT NOT NULL, user_id BIGINT NOT NULL, building_price INT, building_type VARCHAR(50), building_structure VARCHAR(50), building_size DOUBLE, building_date DATE, building_remark VARCHAR(100), is_deleted boolean);
+CREATE TABLE IF NOT EXISTS buildings(id INT PRIMARY KEY AUTO_INCREMENT, project_id INT NOT NULL, user_id INT NOT NULL, building_price BIGINT, building_type VARCHAR(50), building_structure VARCHAR(50), building_size DOUBLE, building_date DATE, building_remark VARCHAR(100), is_deleted boolean);
 
-CREATE TABLE IF NOT EXISTS income_and_expenses(id INT PRIMARY KEY AUTO_INCREMENT,project_id INT NOT NULL, user_id BIGINT NOT NULL, rent INT, maintenance_cost INT, repair_fund INT, management_fee INT, principal INT, interest INT, tax INT, water_bill INT, electric_bill INT, gas_bill INT, fire_insurance INT, other VARCHAR(100), is_deleted boolean);
+CREATE TABLE IF NOT EXISTS income_and_expenses(id INT PRIMARY KEY AUTO_INCREMENT,project_id INT NOT NULL, user_id INT NOT NULL, rent INT, maintenance_cost INT, repair_fund INT, management_fee INT, principal INT, interest INT, tax INT, water_bill INT, electric_bill INT, gas_bill INT, fire_insurance INT, other VARCHAR(100), is_deleted boolean);
 
 CREATE TABLE IF NOT EXISTS users (
-    id BIGINT PRIMARY KEY AUTO_INCREMENT,
+    id INT PRIMARY KEY AUTO_INCREMENT,
     username VARCHAR(50) NOT NULL,
     password VARCHAR(255) NOT NULL,
     email VARCHAR(100) NOT NULL,
